# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Comment:
    properties:
      blog_id:
        description: |-
          ID of the blog post this comment belongs to
          example: 507f1f77bcf86cd799439022
        type: string
      content:
        description: |-
          Content/text of the comment
          example: This is a great post! Thanks for sharing.
        type: string
      created_at:
        description: |-
          Timestamp when the comment was created
          example: 2023-08-10T15:04:05Z07:00
        type: string
      id:
        description: |-
          Unique identifier for the comment
          example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        description: |-
          Timestamp when the comment was last updated
          example: 2023-08-11T16:05:06Z07:00
        type: string
      user_id:
        description: |-
          ID of the user who wrote the comment
          example: 507f1f77bcf86cd799439033
        type: string
      username:
        description: |-
          Username of the commenter
          example: johndoe
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Blog API documentation.
  title: Blog API
  version: "1.0"
paths:
...
  /auth/google/callback:
    get:
      consumes:
      - application/json
      description: Handles OAuth callback, exchanges code for tokens and user info
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OAuth login successful with user info and tokens
          schema:
            additionalProperties: true
            type: object
        "400":
          description: OAuth failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle Google OAuth callback
      tags:
      - auth
...
  /blogs/{blog_id}/comments:
    get:
      description: Retrieve comments with pagination for a specific blog post
      parameters:
      - description: Blog ID
        in: path
        name: blog_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of comments per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get paginated comments for a blog post
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a comment on a blog post. Requires authentication.
      parameters:
      - description: Blog ID
        in: path
        name: blog_id
        required: true
        type: string
      - description: Comment object
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/domain.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a new comment to a blog post
      tags:
      - Comments
...
  /comments/{comment_id}:
    delete:
      description: Delete a comment if the user is the owner or admin
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a comment if the user is the owner
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: Updated comment content
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/domain.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing comment
      tags:
      - Comments
...
swagger: "2.0"